requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch
Django==1.11.5
django-allauth==0.33.0
django-badges==0.2.0
django-nocaptcha-recaptcha==0.0.19
django-reversion==2.0.10
django-reversion-compare==0.8.0
django-updown==1.0.2
djipsum==1.1.4
dnspython==1.15.0
email-validator==1.0.2
Faker==0.8.3
idna==2.6
oauthlib==2.0.3
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.22
